修复用户登陆重复写库的问题，并在问题详情页面添加修改按钮

+++ src/main/java/com/yudabing/community/dto/QuestionDTO.java
@@ -16,7 +16,7 @@ public class QuestionDTO {
     private String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private Long creator;
+    private Integer creator;
     private Integer viewCount;
     private Integer commentCount;
     private Integer likeCount;
